ajout class scoreDealer
